# docs/sidecar/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RUDI Sidecar API",
    "version": "0.1.0",
    "description": "Machine-readable contract for the hardened RUDI sidecar surfaces: health, projects, notes, stable session endpoints, shell, terminal, filesystem, run-groups, and the public run-group WebSocket events.",
    "x-rudi-cli-version": "1.10.12"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:{port}",
      "description": "Local sidecar server",
      "variables": {
        "port": {
          "default": "8100",
          "description": "Dynamic sidecar port written to ~/.rudi/.rudi-lite-port"
        }
      }
    }
  ],
  "security": [
    {
      "RudiTokenAuth": []
    }
  ],
...
  "paths": {
...
    "/terminal/open": {
      "post": {
        "tags": [
          "Terminal"
        ],
        "summary": "Open or reuse an embedded terminal session",
        "description": "Opens a PTY-backed terminal. Requires the optional `@lydell/node-pty` dependency; otherwise the sidecar returns `503`.",
        "operationId": "openTerminal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalOpenRequest"
              },
              "example": {
                "sessionKey": "global",
                "cwd": "/Users/hoff/dev/RUDI",
                "shell": "/bin/zsh",
                "cols": 80,
                "rows": 24
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terminal opened or reused",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TerminalOpenResponse"
                },
                "example": {
                  "ok": true,
                  "sessionKey": "global",
                  "reused": false
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "409": {
            "$ref": "#/components/responses/ConflictError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableError"
          }
        }
      }
    },
    "/terminal/write": {
      "post": {
        "tags": [
          "Terminal"
        ],
        "summary": "Write input to an embedded terminal session",
        "operationId": "writeTerminal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalWriteRequest"
              },
              "example": {
                "sessionKey": "global",
                "data": "ls\\n"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terminal write complete",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/terminal/resize": {
      "post": {
        "tags": [
          "Terminal"
        ],
        "summary": "Resize an embedded terminal session",
        "operationId": "resizeTerminal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalResizeRequest"
              },
              "example": {
                "sessionKey": "global",
                "cols": 120,
                "rows": 30
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terminal resized",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/MissingRequiredFieldError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/terminal/close": {
      "post": {
        "tags": [
          "Terminal"
        ],
        "summary": "Close an embedded terminal session",
        "description": "Idempotent. Closing a nonexistent session still returns `{ ok: true }`.",
        "operationId": "closeTerminal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TerminalSessionKeyRequest"
              },
              "example": {
                "sessionKey": "global"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Terminal closed",
            "headers": {
              "x-rudi-request-id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "RudiTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rudi-token",
        "description": "Sidecar auth token read from ~/.rudi/.rudi-lite-token."
      }
    },
    "headers": {
      "RequestIdHeader": {
        "description": "Per-request correlation ID returned on sidecar responses.",
        "schema": {
          "type": "string"
        }
      }
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Unauthorized",
              "code": "UNAUTHORIZED",
              "requestId": "req_example_123"
            }
          }
        }
      },
...
      "ConflictError": {
        "description": "Conflict",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Conflict",
              "code": "CONFLICT",
              "requestId": "req_example_123"
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Not found",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Not found",
              "code": "NOT_FOUND",
              "requestId": "req_example_123"
            }
          }
        }
      },
...
      "ServiceUnavailableError": {
        "description": "Service unavailable",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "Service unavailable",
              "code": "SERVICE_UNAVAILABLE",
              "requestId": "req_example_123"
            }
          }
        }
      },
      "MissingRequiredFieldError": {
        "description": "Required field missing",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "example": {
              "error": "name required",
              "code": "MISSING_REQUIRED_FIELD",
              "details": {
                "field": "name",
                "location": "body"
              },
              "requestId": "req_example_123"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Missing required field or invalid field value.",
        "headers": {
          "x-rudi-request-id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SidecarError"
            },
            "examples": {
              "missingRequiredField": {
                "value": {
                  "error": "path required",
                  "code": "MISSING_REQUIRED_FIELD",
                  "details": {
                    "field": "path",
                    "location": "body"
                  },
                  "requestId": "req_example_123"
                }
              },
              "invalidPath": {
                "value": {
                  "error": "path must be an absolute filesystem path",
                  "code": "INVALID_FIELD",
                  "details": {
                    "field": "path",
                    "location": "body",
                    "reason": "absolute_path_required"
                  },
                  "requestId": "req_example_123"
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "SidecarError": {
        "type": "object",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          },
          "requestId": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          }
        }
      },
...
      "AbsolutePath": {
        "type": "string",
        "description": "Absolute local filesystem path. Empty, relative, and NUL-containing values are rejected.",
        "examples": [
          "/Users/hoff/dev/RUDI/tmp/example.txt"
        ]
      },
...
      "TerminalShellPath": {
        "type": "string",
        "enum": [
          "/bin/zsh",
          "/bin/bash",
          "/bin/sh"
        ],
        "description": "Allowed interactive shell executable for embedded terminal sessions."
      },
      "TerminalSessionKeyRequest": {
        "type": "object",
        "properties": {
          "sessionKey": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "TerminalOpenRequest": {
        "type": "object",
        "required": [
          "cwd"
        ],
        "properties": {
          "sessionKey": {
            "type": "string"
          },
          "cwd": {
            "$ref": "#/components/schemas/AbsolutePath"
          },
          "shell": {
            "$ref": "#/components/schemas/TerminalShellPath"
          },
          "cols": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000,
            "default": 80
          },
          "rows": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000,
            "default": 24
          }
        },
        "additionalProperties": false
      },
      "TerminalOpenResponse": {
        "type": "object",
        "required": [
          "ok",
          "sessionKey",
          "reused"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "sessionKey": {
            "type": "string"
          },
          "reused": {
            "type": "boolean"
          },
          "buffer": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "additionalProperties": false
      },
      "TerminalWriteRequest": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "sessionKey": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "TerminalResizeRequest": {
        "type": "object",
        "required": [
          "cols",
          "rows"
        ],
        "properties": {
          "sessionKey": {
            "type": "string"
          },
          "cols": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
          },
          "rows": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}